third_party/blink/web_tests/serial/serialPort_readable_closeLocked.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  await port.open({ baudrate: 9600 });
  assert_true(port.readable instanceof ReadableStream);

  const reader = port.readable.getReader();
  await promise_rejects_js(t, TypeError, port.close());

  reader.releaseLock();
  await port.close();
  assert_equals(port.readable, null);
}, 'Port cannot be closed while readable is locked');

</script>
